Reorganize fluent package
