+++ packages/fluent/src/components/ScaleDropdown/ColorChips.tsx
