+++ packages/fluent/src/components/ScaleDropdown/ContinuousBand.tsx
