+++ packages/fluent/src/components/ScaleDropdown/ScaleDropdownItem.tsx
