+++ packages/fluent/src/components/ScaleDropdown/hooks/useSafeDimensions.ts
