+++ packages/fluent/src/components/ScaleDropdown/index.tsx
