+++ packages/fluent/src/components/ScaleDropdown/util.ts
