+++ packages/fluent/src/components/ScaleTypeChoiceGroup/index.tsx
