+++ packages/fluent/src/components/index.ts
@@ -1,0 +1,8 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+export * from './ColorPicker'
+export * from './ColorPickerButton'
+export * from './ScaleDropdown'
+export * from './ScaleTypeChoiceGroup'
