+++ packages/fluent/src/provider/ThematicFluentProvider.tsx
@@ -6,12 +6,11 @@ import { initializeIcons } from '@fluentui/font-icons-mdl2'
 import { ThemeProvider } from '@fluentui/react'
 import { Theme } from '@thematic/core'
 import { ThematicProvider } from '@thematic/react'
-import { useMemo, ReactNode, FC } from 'react'
-import { loadFluentTheme } from './loader'
+import { useMemo, FC } from 'react'
+import { loadFluentTheme } from '../loader'
 
 export interface ThematicFluentProviderProps {
 	theme: Theme
-	children?: ReactNode
 }
 
 initializeIcons()
