+++ packages/fluent/src/theme/index.ts
@@ -2,7 +2,5 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-export * from './components'
-export * from './loader'
-export * from './provider'
-export * from './types'
+export * from './shades'
+export * from './themeJson'
